renames package grouper => groupers
