+++ src/dimcat/steps/groupers/__init__.py
@@ -1,3 +1,5 @@
 import logging
 
+from .base import CustomPieceGrouper
+
 logger = logging.getLogger(__name__)
