+++ src/dimcat/steps/groupers/base.py
