+++ tests/test_pipeline.py
@@ -2,7 +2,7 @@ from dimcat.base import DimcatConfig
 from dimcat.data.resource.dc import PieceIndex
 from dimcat.steps.base import FeatureProcessingStep
 from dimcat.steps.extractors.base import FeatureExtractor
-from dimcat.steps.grouper.base import CustomPieceGrouper
+from dimcat.steps.groupers.base import CustomPieceGrouper
 from dimcat.steps.pipeline.base import Pipeline
